# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Forms Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training-certificates:
    get:
      operationId: listTrainingCertificates
      tags: [training-certificates]
      summary: List all training certificates
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: archived
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of training certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCertificate"
    post:
      operationId: createTrainingCertificate
      tags: [training-certificates]
      summary: Create a training certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingCertificateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingCertificate"
  /training-certificates/{id}:
    get:
      operationId: getTrainingCertificate
      tags: [training-certificates]
      summary: Get a training certificate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingCertificate"
        "404":
          description: Not found
    patch:
      operationId: updateTrainingCertificate
      tags: [training-certificates]
      summary: Update a training certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingCertificateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingCertificate"
    delete:
      operationId: deleteTrainingCertificate
      tags: [training-certificates]
      summary: Archive a training certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Archived
  /training-certificates/{id}/restore:
    patch:
      operationId: restoreTrainingCertificate
      tags: [training-certificates]
      summary: Restore an archived training certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingCertificate"

  # ─── MEDICAL REPORTS ──────────────────────────────────────────────────────────
  /medical-reports:
    get:
      operationId: listMedicalReports
      tags: [medical-reports]
      summary: List all medical reports
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: archived
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of medical reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalReport"
    post:
      operationId: createMedicalReport
      tags: [medical-reports]
      summary: Create a medical report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalReport"
  /medical-reports/{id}:
    get:
      operationId: getMedicalReport
      tags: [medical-reports]
      summary: Get a medical report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalReport"
        "404":
          description: Not found
    patch:
      operationId: updateMedicalReport
      tags: [medical-reports]
      summary: Update a medical report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalReportUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalReport"
    delete:
      operationId: deleteMedicalReport
      tags: [medical-reports]
      summary: Archive a medical report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Archived
  /medical-reports/{id}/restore:
    patch:
      operationId: restoreMedicalReport
      tags: [medical-reports]
      summary: Restore an archived medical report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalReport"

  # ─── INVOICES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TrainingCertificate:
      type: object
      required: [id, certificateNumber, traineeName, gender, department, startDate, endDate, issuedAt, isArchived, createdAt]
      properties:
        id:
          type: integer
        certificateNumber:
          type: string
        traineeName:
          type: string
        gender:
          type: string
          enum: [male, female]
        department:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        trainingOrg:
          type: ["string", "null"]
        issuedAt:
          type: string
        isArchived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    TrainingCertificateInput:
      type: object
      required: [traineeName, gender, department, startDate, endDate]
      properties:
        traineeName:
          type: string
          minLength: 1
        gender:
          type: string
          enum: [male, female]
        department:
          type: string
          minLength: 1
        startDate:
          type: string
        endDate:
          type: string
        trainingOrg:
          type: string
    TrainingCertificateUpdate:
      type: object
      properties:
        traineeName:
          type: string
          minLength: 1
        gender:
          type: string
          enum: [male, female]
        department:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        trainingOrg:
          type: string

    # ─── MEDICAL REPORT ─────────────────────────────────────────────────────────
    MedicalReport:
      type: object
      required: [id, reportNumber, patientName, age, gender, diagnosis, reportText, reportDate, isArchived, createdAt]
      properties:
        id:
          type: integer
        reportNumber:
          type: string
        patientName:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        diagnosis:
          type: string
        reportText:
          type: string
        doctorId:
          type: ["integer", "null"]
        doctorName:
          type: ["string", "null"]
        doctorSpecialty:
          type: ["string", "null"]
        reportDate:
          type: string
        isArchived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    MedicalReportInput:
      type: object
      required: [patientName, age, gender, diagnosis, reportText, reportDate]
      properties:
        patientName:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 0
        gender:
          type: string
          enum: [male, female]
        diagnosis:
          type: string
          minLength: 1
        reportText:
          type: string
          minLength: 1
        doctorId:
          type: integer
        reportDate:
          type: string
    MedicalReportUpdate:
      type: object
      properties:
        patientName:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        diagnosis:
          type: string
        reportText:
          type: string
        doctorId:
          type: integer
        reportDate:
          type: string

    # ─── INVOICE ────────────────────────────────────────────────────────────────
...