# lib/api-zod/src/generated/types/trainingCertificate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Forms Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { TrainingCertificateGender } from './trainingCertificateGender';

export interface TrainingCertificate {
  id: number;
  certificateNumber: string;
  traineeName: string;
  gender: TrainingCertificateGender;
  department: string;
  startDate: string;
  endDate: string;
  /** @nullable */
  trainingOrg?: string | null;
  issuedAt: string;
  isArchived: boolean;
  createdAt: string;
  updatedAt?: string;
}